PR S4.3: decompose moderation_cog into views/moderation/ + cogs/moderation/
Applies the F-3 + Pattern B layout to the moderation subsystem so the
cog file is back to a thin Discord adapter (212 LOC, was 679) and the
PersistentView + modals live where readers expect them.  Same shape
as the just-completed S4.2-followup xp extraction, with the added
Pattern B step required by ModPanelView being a PersistentView.

New layout (mirrors views/mining/ + cogs/mining/):

  cogs/moderation_cog.py            — ModerationCog (8 commands), help-menu
                                      hook, Pattern B re-export of ModPanelView
  cogs/moderation/__init__.py       — package docstring
  cogs/moderation/_helpers.py       — _build_mod_panel_embed,
                                      _can_act_on_interaction
  views/moderation/__init__.py      — re-exports (triggers @register)
  views/moderation/main_panel.py    — @register ModPanelView (Pattern B)
  views/moderation/modals.py        — 7 admin modals

Behavior preserved end-to-end:

  * All 8 text commands (!warn, !timeout, !kick, !ban, !unban,
    !clearwarnings, !modlogs, !modmenu) work unchanged.
  * ModPanelView still has the deliberate collaborative-staff
    interaction_check override (any moderate_members user, not just
    the invoker).
  * Custom_ids (mod:warn, mod:timeout, mod:kick, mod:ban, mod:unban,
    mod:logs, mod:clearwarn) unchanged — restart-safe view binding
    for existing panel_anchors rows preserved.
  * build_help_menu_view still returns (embed, ModPanelView()) so the
    help-cog direct-navigation flow lands on the live panel.
  * @register fires at cog-load time because cogs/moderation_cog.py
    does `from views.moderation import ModPanelView` at module level.
    Verified: _REGISTRY["moderation"] points to the new
    views.moderation.main_panel.ModPanelView class.

Identity contract intact:

  * SUBSYSTEMS["moderation"] keys unchanged.
  * persistent_views._REGISTRY["moderation"] populated at the same
    moment as before (module import of cogs.moderation_cog triggers
    the chain).
  * bot.commands unchanged (all 8 commands still on ModerationCog).
  * panel_anchors rows store subsystem="moderation" — unchanged.

Validation:

  * pytest tests/unit/ → 837 passed, 0 failed
  * Identity contract STRICT mode green (21 contract tests)
  * test_moderation_build_help_menu_view_returns_embed_and_view passes
    against the new layout without modification
  * INV-A..INV-N green
  * black, isort, ruff, mypy all clean over the whole disbot/ tree
  * Manual import smoke test confirms ModPanelView is the same class
    in both `cogs.moderation_cog.ModPanelView` (Pattern B re-export)
    and `views.moderation.main_panel.ModPanelView` (definition site).

Stage A2 of Phase A complete.  Next: A3 (rps_tournament_cog
decomposition, depends only on A1 which is already in this branch).
