+++ disbot/cogs/moderation_cog.py
@@ -7,485 +7,18 @@ import discord
 from discord import Member
 from discord.ext import commands
 
+from cogs.moderation._helpers import _build_mod_panel_embed
 from core.runtime import panel_manager
-from core.runtime.component_registry import stats_block
-from core.runtime.interaction_helpers import safe_defer
-from core.runtime.persistent_views import PersistentView, register
 from utils import db
-from utils.helpers import _parse_member
 from utils.settings_keys import WARN_THRESHOLD, WARN_TIMEOUT_MINS
 from utils.ui_constants import MOD_COLOR
 
+# Pattern B re-export: importing this triggers @register on ModPanelView
+# so the persistent-view registry is populated before on_ready runs
+# restore_anchors.  See docs/architecture.md §"PersistentView placement".
+from views.moderation import ModPanelView  # noqa: F401 — re-exported
+
 logger = logging.getLogger("bot")
-
-
-def _build_mod_panel_embed() -> discord.Embed:
-    return stats_block(
-        "🔨 Moderation Panel",
-        [
-            ("⚠️ Warn", "Issue a warning (auto-timeout at 3)", True),
-            ("⏳ Timeout", "Temporarily mute for N minutes", True),
-            ("👢 Kick", "Remove from server", True),
-            ("🚫 Ban", "Permanently ban", True),
-            ("✅ Unban", "Lift a ban by user ID", True),
-            ("📋 Mod Logs", "View moderation history", True),
-            ("⬛ Clear Warnings", "Reset warning count", True),
-        ],
-        MOD_COLOR,
-        description=(
-            "Click a button to take a moderation action.\n"
-            "You'll be prompted to enter the user and reason."
-        ),
-        footer="Any staff member with Moderate Members permission may use this panel.",
-    )
-
-
-def _can_act_on_interaction(
-    interaction: discord.Interaction,
-    member: Member,
-) -> str | None:
-    """Return an error string if the actor cannot moderate *member*, else None."""
-    if member == interaction.guild.owner:
-        return "❌ You cannot perform this action on the server owner."
-    actor = interaction.guild.get_member(interaction.user.id)
-    if actor and member.top_role >= actor.top_role:
-        return (
-            "❌ You cannot perform this action on someone with an equal or higher role."
-        )
-    if member.top_role >= interaction.guild.me.top_role:
-        return (
-            "❌ I cannot perform this action — that member has a higher role than me."
-        )
-    return None
-
-
-# ---------------------------------------------------------------------------
-# Modals
-# ---------------------------------------------------------------------------
-
-
-class _WarnModal(discord.ui.Modal, title="Warn Member"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-    reason_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="Reason",
-        style=discord.TextStyle.paragraph,
-        required=False,
-        placeholder="No reason provided",
-        max_length=500,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        reason = self.reason_input.value or "No reason provided"
-        err = _can_act_on_interaction(interaction, member)
-        if err:
-            await interaction.response.send_message(err, ephemeral=True)
-            return
-        threshold = int(await db.get_setting(interaction.guild_id, WARN_THRESHOLD, "3"))
-        timeout_minutes = int(
-            await db.get_setting(interaction.guild_id, WARN_TIMEOUT_MINS, "10"),
-        )
-        count = await db.add_warning(member.id, interaction.guild_id)
-        await interaction.response.send_message(
-            f"⚠️ {member.mention} warned ({count}/{threshold}). Reason: {reason}",
-            ephemeral=False,
-        )
-        await db.log_mod_action(
-            interaction.guild_id,
-            "warn",
-            member.id,
-            interaction.user.id,
-            reason,
-        )
-        if count >= threshold:
-            try:
-                until = discord.utils.utcnow() + timedelta(minutes=timeout_minutes)
-                await member.timeout(until, reason=f"{threshold} warnings reached.")
-                await interaction.followup.send(
-                    f"⏳ {member.mention} timed out for {timeout_minutes} minutes "
-                    f"({threshold} warnings).",
-                )
-                await db.clear_warnings(member.id, interaction.guild_id)
-            except discord.Forbidden:
-                await interaction.followup.send(
-                    f"⚠️ {threshold} warnings reached but I lack permission to timeout.",
-                    ephemeral=True,
-                )
-
-
-class _TimeoutModal(discord.ui.Modal, title="Timeout Member"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-    duration_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="Duration (minutes)",
-        placeholder="e.g. 30",
-        max_length=10,
-    )
-    reason_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="Reason",
-        style=discord.TextStyle.paragraph,
-        required=False,
-        placeholder="No reason provided",
-        max_length=500,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        if not self.duration_input.value.isdigit():
-            await interaction.response.send_message(
-                "❌ Duration must be a whole number of minutes.",
-                ephemeral=True,
-            )
-            return
-        duration = int(self.duration_input.value)
-        reason = self.reason_input.value or "No reason provided"
-        err = _can_act_on_interaction(interaction, member)
-        if err:
-            await interaction.response.send_message(err, ephemeral=True)
-            return
-        try:
-            until = discord.utils.utcnow() + timedelta(minutes=duration)
-            await member.timeout(until, reason=reason)
-            await interaction.response.send_message(
-                f"⏳ {member.mention} timed out for {duration} minute(s).",
-            )
-            await db.log_mod_action(
-                interaction.guild_id,
-                "timeout",
-                member.id,
-                interaction.user.id,
-                f"{duration}m: {reason}",
-            )
-        except discord.Forbidden:
-            await interaction.response.send_message(
-                "❌ No permission to timeout that user.",
-                ephemeral=True,
-            )
-
-
-class _KickModal(discord.ui.Modal, title="Kick Member"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-    reason_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="Reason",
-        style=discord.TextStyle.paragraph,
-        required=False,
-        placeholder="No reason provided",
-        max_length=500,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        reason = self.reason_input.value or "No reason provided"
-        err = _can_act_on_interaction(interaction, member)
-        if err:
-            await interaction.response.send_message(err, ephemeral=True)
-            return
-        try:
-            await member.kick(reason=reason)
-            await interaction.response.send_message(
-                f"👢 {member.mention} kicked. Reason: {reason}",
-            )
-            await db.log_mod_action(
-                interaction.guild_id,
-                "kick",
-                member.id,
-                interaction.user.id,
-                reason,
-            )
-        except discord.Forbidden:
-            await interaction.response.send_message(
-                "❌ No permission to kick that user.",
-                ephemeral=True,
-            )
-
-
-class _BanModal(discord.ui.Modal, title="Ban Member"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-    reason_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="Reason",
-        style=discord.TextStyle.paragraph,
-        required=False,
-        placeholder="No reason provided",
-        max_length=500,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        reason = self.reason_input.value or "No reason provided"
-        err = _can_act_on_interaction(interaction, member)
-        if err:
-            await interaction.response.send_message(err, ephemeral=True)
-            return
-        try:
-            await member.ban(reason=reason)
-            await interaction.response.send_message(
-                f"🚫 {member.mention} banned. Reason: {reason}",
-            )
-            await db.log_mod_action(
-                interaction.guild_id,
-                "ban",
-                member.id,
-                interaction.user.id,
-                reason,
-            )
-        except discord.Forbidden:
-            await interaction.response.send_message(
-                "❌ No permission to ban that user.",
-                ephemeral=True,
-            )
-
-
-class _UnbanModal(discord.ui.Modal, title="Unban Member"):  # type: ignore[call-arg]
-    user_id_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User ID",
-        placeholder="Right-click user → Copy ID",
-        max_length=20,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        if not await safe_defer(interaction, ephemeral=True):
-            return
-        raw = self.user_id_input.value.strip()
-        if not raw.isdigit():
-            await interaction.followup.send("❌ Please enter a valid numeric user ID.")
-            return
-        user_id = int(raw)
-        try:
-            user = await interaction.client.fetch_user(user_id)
-        except discord.NotFound:
-            await interaction.followup.send(f"❌ No user found with ID `{user_id}`.")
-            return
-        except discord.HTTPException as e:
-            await interaction.followup.send(f"❌ Failed to fetch user: {e}")
-            return
-        try:
-            await interaction.guild.unban(user)
-            await interaction.followup.send(f"✅ {user.mention} unbanned.")
-            await db.log_mod_action(
-                interaction.guild_id,
-                "unban",
-                user.id,
-                interaction.user.id,
-                "",
-            )
-        except discord.NotFound:
-            await interaction.followup.send(f"❌ User `{user_id}` is not banned.")
-        except discord.Forbidden:
-            await interaction.followup.send("❌ No permission to unban.")
-        except discord.HTTPException as e:
-            await interaction.followup.send(f"❌ Failed to unban: {e}")
-
-
-class _ModLogsModal(discord.ui.Modal, title="View Mod Logs"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        logs = await db.get_mod_logs(member.id, interaction.guild_id, limit=10)
-        embed = discord.Embed(
-            title=f"📋 Mod Logs — {member.display_name}",
-            color=MOD_COLOR,
-        )
-        if not logs:
-            embed.description = "No moderation history found."
-        else:
-            for entry in logs:
-                embed.add_field(
-                    name=f"{entry['action'].upper()} — {entry['timestamp']}",
-                    value=f"By <@{entry['moderator_id']}> | {entry['reason']}",
-                    inline=False,
-                )
-        await interaction.response.send_message(embed=embed, ephemeral=True)
-
-
-class _ClearWarningsModal(discord.ui.Modal, title="Clear Warnings"):  # type: ignore[call-arg]
-    member_input = discord.ui.TextInput(  # type: ignore[var-annotated]
-        label="User (mention, ID, or name)",
-        max_length=100,
-    )
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    async def on_submit(self, interaction: discord.Interaction):
-        member = _parse_member(interaction.guild, self.member_input.value)
-        if not member:
-            await interaction.response.send_message(
-                "❌ Member not found.",
-                ephemeral=True,
-            )
-            return
-        await db.clear_warnings(member.id, interaction.guild_id)
-        await db.log_mod_action(
-            interaction.guild_id,
-            "clearwarnings",
-            member.id,
-            interaction.user.id,
-            "",
-        )
-        await interaction.response.send_message(
-            f"✅ Warnings cleared for {member.mention}.",
-            ephemeral=True,
-        )
-
-
-# ---------------------------------------------------------------------------
-# Moderation action panel view
-# ---------------------------------------------------------------------------
-
-
-@register
-class ModPanelView(PersistentView):
-    """Persistent moderation dashboard — any staff member with moderate_members can use it.
-
-    Intentional exception to invoker-only ownership: moderation panels are
-    collaborative staff tools.  Any moderator in the guild may interact.
-    """
-
-    SUBSYSTEM = "moderation"
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if not interaction.user.guild_permissions.moderate_members:  # type: ignore[union-attr]
-            await interaction.response.send_message(
-                "❌ You need Moderate Members permission.",
-                ephemeral=True,
-            )
-            return False
-        return True
-
-    @discord.ui.button(
-        label="⚠️ Warn",
-        style=discord.ButtonStyle.primary,
-        row=0,
-        custom_id="mod:warn",
-    )
-    async def warn_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_WarnModal())
-
-    @discord.ui.button(
-        label="⏳ Timeout",
-        style=discord.ButtonStyle.primary,
-        row=0,
-        custom_id="mod:timeout",
-    )
-    async def timeout_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_TimeoutModal())
-
-    @discord.ui.button(
-        label="👢 Kick",
-        style=discord.ButtonStyle.danger,
-        row=0,
-        custom_id="mod:kick",
-    )
-    async def kick_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_KickModal())
-
-    @discord.ui.button(
-        label="🚫 Ban",
-        style=discord.ButtonStyle.danger,
-        row=1,
-        custom_id="mod:ban",
-    )
-    async def ban_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_BanModal())
-
-    @discord.ui.button(
-        label="✅ Unban",
-        style=discord.ButtonStyle.success,
-        row=1,
-        custom_id="mod:unban",
-    )
-    async def unban_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_UnbanModal())
-
-    @discord.ui.button(
-        label="📋 Mod Logs",
-        style=discord.ButtonStyle.grey,
-        row=1,
-        custom_id="mod:logs",
-    )
-    async def modlogs_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        await interaction.response.send_modal(_ModLogsModal())
-
-    @discord.ui.button(
-        label="⬛ Clear Warnings",
-        style=discord.ButtonStyle.grey,
-        row=2,
-        custom_id="mod:clearwarn",
-    )
-    async def clearwarn_btn(
-        self,
-        interaction: discord.Interaction,
-        _: discord.ui.Button,
-    ):
-        await interaction.response.send_modal(_ClearWarningsModal())
-
-
-# ---------------------------------------------------------------------------
-# Cog
-# ---------------------------------------------------------------------------
 
 
 class ModerationCog(commands.Cog):
